app/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.4.0
requests==2.32.2